Usando los roles en la api

+++ laravel-full/routes/api.php
@@ -26,46 +26,36 @@ use App\Http\Controllers\AuthController;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
-    Route::get('generos',[GeneroController::class,'index']);
-    Route::post('generos',[GeneroController::class,'store']);
-    Route::get('generos/{id}',[GeneroController::class,'show']);
-    Route::put('generos/{id}',[GeneroController::class,'update']);
-    Route::delete('generos/{id}',[GeneroController::class,'destroy']);
-    
-    
-    Route::get('usuarios',[UsuarioController::class,'index']);
-    Route::get('usuarios/{id}',[UsuarioController::class,'show']);
-    Route::put('usuarios/{id}',[UsuarioController::class,'update']);
-    Route::post('usuarios',[UsuarioController::class,'store']);
-    
-    
-    
-    Route::post('proveedores',[ProveedorController::class,'store']);
-    
-    
-    //Roles
-    Route::get('roles',[RolController::class,'index']);
-    Route::post('roles',[RolController::class,'store']);
-    Route::get('roles/{id}',[RolController::class,'show']);
-    Route::put('roles/{id}',[RolController::class,'update']);
-    Route::delete('roles/{id}',[RolController::class,'destroy']);
-    //Permisos
-    Route::get('permisos',[PermisoController::class,'index']);
-    
-    //RolesPermisos
-    Route::get('rolespermisos',[RolePermisoController::class,'index']);
-    
-    //ProveedorRol
-    Route::get('proveedoresRoles',[ProveedorRolController::class,'index']);
-    Route::post('proveedoresRoles',[ProveedorRolController::class,'store']);
-    //Route::put('proveedoresRoles/{id}',[ProveedorRolController::class,'update']);
-    
-    //ProveedorAuth
-    Route::post('proveedoresAuth',[ProveedorAuth::class,'store']);
-Route::group(['middleware' => ['jwt.verify']], function() {
-//Generos
 
-});
+//generos
+Route::get('generos',[GeneroController::class,'index'])->middleware('can:super-admin,moderador,escritor');
+Route::post('generos',[GeneroController::class,'store'])->middleware('can:super-admin');
+Route::get('generos/{id}',[GeneroController::class,'show'])->middleware('can:super-admin,moderador,escritor');
+Route::delete('generos/{id}',[GeneroController::class,'destroy'])->middleware('can:super-admin');
+Route::put('generos/{id}',[GeneroController::class,'update'])->middleware('can:super-admin');
+
+
+//Roles
+Route::get('roles',[RolController::class,'index'])->middleware('can:super-admin');
+Route::post('roles',[RolController::class,'store'])->middleware('can:super-admin');
+Route::get('roles/{id}',[RolController::class,'show'])->middleware('can:super-admin');
+Route::put('roles/{id}',[RolController::class,'update'])->middleware('can:super-admin');
+Route::delete('roles/{id}',[RolController::class,'destroy'])->middleware('can:super-admin');
+//Permisos
+Route::get('permisos',[PermisoController::class,'index'])->middleware('can:super-admin');
+
+//RolesPermisos
+Route::get('rolespermisos',[RolePermisoController::class,'index'])->middleware('can:super-admin');
+
+//ProveedorRol
+Route::get('proveedoresRoles',[ProveedorRolController::class,'index'])->middleware('can:super-admin');
+Route::post('proveedoresRoles',[ProveedorRolController::class,'store'])->middleware('can:super-admin');
+//Route::put('proveedoresRoles/{id}',[ProveedorRolController::class,'update']);
+
+
+//Proveedor 
+Route::post('proveedores',[ProveedorController::class,'store'])->middleware('can:super-admin');
+
 
 //AUHT
 Route::post('auth/loginProveedor',[AuthController::class,'loginProveedor']);
